1699 - numbers can now be ints or floats

+++ 011.cc
@@ -128,7 +128,7 @@ COMPILE_FN(set_cdr, compiledFn_set_cdr, "($cons $val)",
 COMPILE_FN(len, compiledFn_len, "($x)",
   Cell* x = lookup("$x");
   if (isString(x))
-    return mkref(newNum(toString(x).length()));
+    return mkref(newNum((int)toString(x).length()));
 
   int ans = 0;
   for (; x != nil; x=cdr(x))
